pyproject.toml
[project]
name = "slouchless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
# Do not edit this section manually; use uv add <package> to add dependencies.
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pystray>=0.19.5",
    "ray>=2.53.0",
    "tk>=0.1.0",
    "vllm>=0.13.0",
]
